requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.9.0
python-multipart>=0.0.7
numpy>=2.2.0
python-dotenv>=1.0.0
sib-api-v3-sdk>=7.0.0